compartmentalized and setup folders correctly based on community accepted structure
