+++ jet_tracking/gui/widgets/basicWidgets.py
@@ -1,0 +1,197 @@
+from PyQt5 import QtCore
+from PyQt5.QtCore import pyqtSignal, QParallelAnimationGroup, QPropertyAnimation, Qt, QAbstractAnimation
+from PyQt5.QtGui import QDoubleValidator
+from PyQt5.QtWidgets import QGraphicsView, QGraphicsScene, QComboBox, QPushButton, QLineEdit, QToolButton, QWidget, \
+    QScrollArea, QSizePolicy, QVBoxLayout, QLabel, QMessageBox
+
+
+class GraphicsView(QGraphicsView):
+    def __init__(self, parent=None):
+        super(GraphicsView, self).__init__(parent)
+        self.setMouseTracking(True)
+
+
+class GraphicsScene(QGraphicsScene):
+    def __init__(self, parent=None):
+        super(GraphicsScene, self).__init__(parent)
+
+
+class ComboBox(QComboBox):
+    def __init__(self, parent=None):
+        super(ComboBox, self).__init__(parent)
+
+
+class PushButton(QPushButton):
+    def __init__(self, parent=None):
+        super(PushButton, self).__init__(parent)
+
+
+class LineEdit(QLineEdit):
+    checkVal = pyqtSignal(float)
+
+    def __init__(self, *args, **kwargs):
+        super(LineEdit, self).__init__(*args, **kwargs)
+        self.validator = QDoubleValidator()
+        self.setValidator(self.validator)
+        self.textChanged.connect(self.new_text)
+        self.returnPressed.connect(self.check_validator)
+        self.ntext = self.text()
+
+    def new_text(self, text):
+        if self.hasAcceptableInput():
+            self.ntext = text
+
+    def keyPressEvent(self, event):
+        super().keyPressEvent(event)
+        if event.key() == Qt.Key_Return and not self.hasAcceptableInput():
+            self.check_validator()
+
+    def check_validator(self):
+        try:
+            if float(self.text()) > self.validator.top():
+                self.setText(str(self.validator.top()))
+            elif float(self.text()) < self.validator.bottom():
+                self.setText(str(self.validator.bottom()))
+        except:
+            mssg = QMessageBox.about(self, "Error", "Input can only be a number")
+            self.setText(self.ntext)
+        self.checkVal.emit(float(self.ntext))
+
+    def valRange(self, x1, x2):
+        self.validator.setRange(x1, x2)
+        self.validator.setDecimals(6)
+        self.validator.setNotation(QDoubleValidator.StandardNotation)
+
+
+class CollapsibleBox(QWidget):
+    def __init__(self, title="", parent=None):
+        super(CollapsibleBox, self).__init__(parent)
+        self.toggle_button = QToolButton(
+            text=title, checkable=True, checked=False
+        )
+        self.toggle_button.setStyleSheet("QToolButton {border: none;\
+                border: 1px solid #FF17365D;\
+                border-top-left-radius: 15px;\
+                border-top-right-radius: 15px;\
+                background-color: #FF17365D;\
+                padding: 5px 0px;\
+                color: rgb(255, 255, 255);\
+                max-height: 30px;\
+                font-size: 14px;\
+            }\
+            QToolButton:hover {\
+                background-color: lightgreen;\
+                color: black;\
+            }")
+        self.toggle_button.setToolButtonStyle(
+            Qt.ToolButtonTextBesideIcon
+        )
+        self.toggle_button.setArrowType(Qt.RightArrow)
+        self.toggle_button.pressed.connect(self.on_pressed)
+
+        self.toggle_animation = QParallelAnimationGroup(self)
+
+        self.content_area = QScrollArea(
+            maximumHeight=0, minimumHeight=0
+        )
+        self.content_area.setSizePolicy(
+            QSizePolicy.Expanding, QSizePolicy.Fixed
+        )
+
+        lay = QVBoxLayout(self)
+        lay.setSpacing(0)
+        lay.setContentsMargins(0, 0, 0, 0)
+        lay.addWidget(self.toggle_button)
+        lay.addWidget(self.content_area)
+
+        self.toggle_animation.addAnimation(
+            QPropertyAnimation(self, b"minimumHeight")
+        )
+        self.toggle_animation.addAnimation(
+            QPropertyAnimation(self, b"maximumHeight")
+        )
+        self.toggle_animation.addAnimation(
+            QPropertyAnimation(self.content_area, b"maximumHeight")
+        )
+
+    @QtCore.pyqtSlot()
+    def on_pressed(self):
+        checked = self.toggle_button.isChecked()
+        self.toggle_button.setArrowType(
+            Qt.DownArrow if not checked else Qt.RightArrow
+        )
+        self.toggle_animation.setDirection(
+            QAbstractAnimation.Forward
+            if not checked
+            else QAbstractAnimation.Backward
+        )
+        self.toggle_animation.start()
+
+    def clear_layout(self, layout):
+        try:
+            for i in reversed(range(layout.count())):
+                widgetToRemove = layout.itemAt(i).widget()
+                layout.removeWidget(widgetToRemove)
+                widgetToRemove.setPArent(None)
+        except AttributeError:
+            pass
+
+    def setContentLayout(self, layout):
+        lay = self.content_area.layout()
+        self.clear_layout(lay)
+        self.content_area.setLayout(layout)
+        collapsed_height = (
+                self.sizeHint().height() - self.content_area.maximumHeight()
+        )
+        content_height = layout.sizeHint().height()
+        for i in range(self.toggle_animation.animationCount()):
+            animation = self.toggle_animation.animationAt(i)
+            animation.setDuration(500)
+            animation.setStartValue(collapsed_height)
+            animation.setEndValue(collapsed_height + content_height)
+
+        content_animation = self.toggle_animation.animationAt(
+            self.toggle_animation.animationCount() - 1
+        )
+        content_animation.setDuration(500)
+        content_animation.setStartValue(0)
+        content_animation.setEndValue(content_height)
+
+
+class Label(QLabel):
+    def __init__(self, parent=None):
+        super(Label, self).__init__(parent)
+
+    def setTitleStylesheet(self):
+        self.setStyleSheet("\
+                qproperty-alignment: AlignCenter;\
+                border: 1px solid #FF17365D;\
+                border-top-left-radius: 15px;\
+                border-top-right-radius: 15px;\
+                background-color: #FF17365D;\
+                padding: 5px 0px;\
+                color: rgb(255, 255, 255);\
+                max-height: 35px;\
+                font-size: 14px;\
+            ")
+
+    def setSubtitleStyleSheet(self):
+        self.setStyleSheet("\
+                qproperty-alignment: AlignCenter;\
+                border: 1px solid #FF17365D;\
+                background-color: #FF17365D;\
+                padding: 5px 0px;\
+                color: rgb(255, 255, 255);\
+                font-size: 12px;\
+            ")
+
+    def setTrackingStylesheet(self):
+        # this will change based on the status given back
+        self.setStyleSheet("\
+                qproperty-alignment: AlignCenter;\
+                border: 1px solid #FF17365D;\
+                background-color: red;\
+                padding: 5px 0px;\
+                color: rgb(255, 255, 255);\
+                font-size: 12px;\
+            ")
