--- jet_tracking/psana/__init__.py
@@ -1,1 +1,0 @@
-from .psana2epics import *  # noqa
