--- jet_tracking/psana/shmem_notes
@@ -1,85 +1,0 @@
-Notes for auto jet tracking
-
-See HXR Jet Tracking google drive for project notes:
-- https://drive.google.com/drive/folders/1TJdb9qK85oxeFpoPQ5m080HZw-Dptm7e
-
-Shared memory application analysis based on GDVN for cxilr6716:
-- see also: /reg/neh/operator/cxiopr/experiments/cxilr6716/src/README
-
-pvNotepad pvs hosted by the ioc-mfx-cspadPV ioc:
-/reg/g/pcds/package/epics/3.14-dev/ioc/mfx/pvNotepad/current/ioc-mfx-cspadPV.cfg
-
------------------------
-Running psana with output to shared memory
-
-psana2epics.sh example will output the sum of the CSPAD intensity to a PV.
-
-e.g., for start of Run 17 we are using SC2
-
-[koglin@daq-cxi-mon01 ~ 09:34:39] /reg/g/pcds/pyps/apps/hutch-python/cxi/cxi/jet_tracking/psana/psana2epics.sh --alias='DsaCsPad' --pvbase='CXI:SC2:DIFFRACT'
-Initializing args Namespace(alias='DsaCsPad', calc_period=False, calc_streak=False, exp='cxilr6716', instrument='cxi', nevents=10, plot=False, pvbase='CXI:SC2:DIFFRACT', run=None)
-Initializing epics PVs
-Events = 240.0
-... done
-Loading DataSource for shared memory with expiment cxilr6716
-No data source specified, so assume this is shared memory.
-setting calibDir cxilr6716 /reg/d/psdm/cxi/cxilr6716/calib
-Opened queue /PdsFromMonitorDiscovery_psana (9)
-Opening shared memory /PdsMonitorSharedMemory_psana of size 0x4c800000 (0x22 * 0x2400000)
-Shared memory at 0x7f7c473dc000
-Opened queue /PdsToMonitorEvQueue_psana_0 (21)
-Opened queue /PdsToMonitorEvQueue_psana_1 (22)
---------------------------------------------------------------------------------
-Source Information:
-------------------
-*Detectors in group 0 are "BLD" data recorded at 120 Hz on event code 40
-*Detectors listed as Controls are controls devices with unknown event code (but likely 40).
-
-Alias                Grp   Description Code Pol   Delay [s]   Width [s] Source
-------------------------------------------------------------------------------------------------------
-CXI_DG2_BMMON          1                                                BldInfo(CXI-DG2-BMMON)
-DsaCsPad               1        120 Hz   40 Neg 0.000790723 0.000040000 DetInfo(CxiDs2.0:Cspad.0)
-EBeam                  1                                                BldInfo(EBeam)
-FEEGasDetEnergy        1                                                BldInfo(FEEGasDetEnergy)
-HFX_DG2_BMMON          1                                                BldInfo(HFX-DG2-BMMON)
-HX2_SB1_BMMON          1                                                BldInfo(HX2-SB1-BMMON)
-KbEncoder              1        120 Hz   40 Pos 0.000490723 0.000010000 DetInfo(CxiEndstation.0:USDUSB.0)
-PhaseCavity            1                                                BldInfo(PhaseCavity)
-Sc2Inline              1        120 Hz   40 Pos 0.000790723 0.007000000 DetInfo(CxiSc2.0:Opal1000.0)
-XCS_SB1_BMMON          1                                                BldInfo(XCS-SB1-BMMON)
---------------------------------------------------------------------------------
-Event Code Information:
-------------------
-code     group type         description
-----------------------------------------------------------
-      40    1 Readout      120 Hz
-
-Load time: 11.4506380558 sec
-... done
---------------------------------------------------------------------------
-
-
-
-For CXI, monitor PV from cxi control room machine
-
-cxi-daq:cxi> camonitor CXI:SC2:DIFFRACT:TOTAL_ADU
-CXI:SC2:DIFFRACT:TOTAL_ADU     2018-09-06 21:35:01.535993 5.33097e+06
-CXI:SC2:DIFFRACT:TOTAL_ADU     2018-09-06 21:35:01.579956 4.1387e+06
-CXI:SC2:DIFFRACT:TOTAL_ADU     2018-09-06 21:35:01.591779 4.3779e+06
-CXI:SC2:DIFFRACT:TOTAL_ADU     2018-09-06 21:35:01.603362 4.53231e+06
-CXI:SC2:DIFFRACT:TOTAL_ADU     2018-09-06 21:35:01.614737 3.38e+06
-...
-
-cxi-daq:cxi> caEventRate CXI:SC2:DIFFRACT:TOTAL_ADU.VAL
-Connecting to CA Channel "CXI:SC2:DIFFRACT:TOTAL_ADU.VAL" 1 times. done(0.028546 sec).
-Subscribing 1 times. done(0.000059 sec).
-Waiting for initial value events. done(0.010099 sec).
-CA Event Rate (Hz): current 40.9977 mean 40.9977 std dev 0
-CA Event Rate (Hz): current 39.4983 mean 40.248 std dev 0.749664
-CA Event Rate (Hz): current 38.4994 mean 39.6651 std dev 1.02671
-...
-
-
-For mfx on daq-mfx-mon02:
-
-daq-mfx-mon02> /reg/g/pcds/pyps/apps/hutch-python/mfx/mfx/jet_tracking/psana/psana2epics.sh --alias='Rayonix' --pvbase='MFX:DET:DIFFRACT' --exp='mfxlt3017'
